Chown the build artifacts to the user invoking the script
The script must be run as root due to debootstrap and chroot, but the
built artifacts should be owned by the user invoking the script. When
running the script with sudo the SUDO_UID and SUDO_GID variables are set
to the user which invoked sudo, and we can chown the artifacts to that
user.