Workaround sporadic "Timed out waiting for device" on boot
The device with the "config" ext4 filesystem isn't always detected on
boot, the issue sounds similar to this[1] upstream issue, and may have
already been fixed (the OS is running systemd v247.3, the newest is
v251.2). As it may have been solved in newer verions of systemd, a
workaround is implemented instead of investigating it further. In this
case, reboot the system if the device isn't detected.

[1] https://github.com/systemd/systemd/issues/13775